# Refa: revise the implementation of LightRAG and enable response caching