--- a/graphrag/general/extractor.py
+++ b/graphrag/general/extractor.py
@@ -1,5 +1,5 @@
 #
-#  Copyright 2024 The InfiniFlow Authors. All Rights Reserved.
+#  Copyright 2025 The InfiniFlow Authors. All Rights Reserved.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -15,16 +15,26 @@
 #
 import logging
 import re
-from collections import defaultdict, Counter
+from collections import Counter, defaultdict
 from copy import deepcopy
 from typing import Callable
-import trio
+
 import networkx as nx
+import trio
 
 from api.utils.api_utils import timeout
 from graphrag.general.graph_prompt import SUMMARIZE_DESCRIPTIONS_PROMPT
-from graphrag.utils import get_llm_cache, set_llm_cache, handle_single_entity_extraction, \
-    handle_single_relationship_extraction, split_string_by_multi_markers, flat_uniq_list, chat_limiter, get_from_to, GraphChange
+from graphrag.utils import (
+    GraphChange,
+    chat_limiter,
+    flat_uniq_list,
+    get_from_to,
+    get_llm_cache,
+    handle_single_entity_extraction,
+    handle_single_relationship_extraction,
+    set_llm_cache,
+    split_string_by_multi_markers,
+)
 from rag.llm.chat_model import Base as CompletionLLM
 from rag.prompts import message_fit_in
 from rag.utils import truncate
@@ -47,7 +57,7 @@ class Extractor:
         self._language = language
         self._entity_types = entity_types or DEFAULT_ENTITY_TYPES
 
-    @timeout(60*20)
+    @timeout(60 * 20)
     def _chat(self, system, history, gen_conf={}):
         hist = deepcopy(history)
         conf = deepcopy(gen_conf)
@@ -75,31 +85,19 @@ class Extractor:
         maybe_edges = defaultdict(list)
         ent_types = [t.lower() for t in self._entity_types]
         for record in records:
-            record_attributes = split_string_by_multi_markers(
-                record, [tuple_delimiter]
-            )
+            record_attributes = split_string_by_multi_markers(record, [tuple_delimiter])
 
-            if_entities = handle_single_entity_extraction(
-                record_attributes, chunk_key
-            )
+            if_entities = handle_single_entity_extraction(record_attributes, chunk_key)
             if if_entities is not None and if_entities.get("entity_type", "unknown").lower() in ent_types:
                 maybe_nodes[if_entities["entity_name"]].append(if_entities)
                 continue
 
-            if_relation = handle_single_relationship_extraction(
-                record_attributes, chunk_key
-            )
+            if_relation = handle_single_relationship_extraction(record_attributes, chunk_key)
             if if_relation is not None:
-                maybe_edges[(if_relation["src_id"], if_relation["tgt_id"])].append(
-                    if_relation
-                )
+                maybe_edges[(if_relation["src_id"], if_relation["tgt_id"])].append(if_relation)
         return dict(maybe_nodes), dict(maybe_edges)
 
-    async def __call__(
-        self, doc_id: str, chunks: list[str],
-            callback: Callable | None = None
-    ):
-
+    async def __call__(self, doc_id: str, chunks: list[str], callback: Callable | None = None):
         self.callback = callback
         start_ts = trio.current_time()
 
@@ -107,7 +105,7 @@ class Extractor:
             out_results = []
             limiter = trio.Semaphore(max_concurrency)
 
-            async def worker(chunk_key_dp: tuple[str, str], idx:int, total:int):
+            async def worker(chunk_key_dp: tuple[str, str], idx: int, total: int):
                 async with limiter:
                     await self._process_single_content(chunk_key_dp, idx, total, out_results)
 
@@ -130,7 +128,7 @@ class Extractor:
             sum_token_count += token_count
         now = trio.current_time()
         if callback:
-            callback(msg = f"Entities and relationships extraction done, {len(maybe_nodes)} nodes, {len(maybe_edges)} edges, {sum_token_count} tokens, {now-start_ts:.2f}s.")
+            callback(msg=f"Entities and relationships extraction done, {len(maybe_nodes)} nodes, {len(maybe_edges)} edges, {sum_token_count} tokens, {now - start_ts:.2f}s.")
         start_ts = now
         logging.info("Entities merging...")
         all_entities_data = []
@@ -139,7 +137,7 @@ class Extractor:
                 nursery.start_soon(self._merge_nodes, en_nm, ents, all_entities_data)
         now = trio.current_time()
         if callback:
-            callback(msg = f"Entities merging done, {now-start_ts:.2f}s.")
+            callback(msg=f"Entities merging done, {now - start_ts:.2f}s.")
 
         start_ts = now
         logging.info("Relationships merging...")
@@ -149,12 +147,10 @@ class Extractor:
                 nursery.start_soon(self._merge_edges, src, tgt, rels, all_relationships_data)
         now = trio.current_time()
         if callback:
-            callback(msg = f"Relationships merging done, {now-start_ts:.2f}s.")
+            callback(msg=f"Relationships merging done, {now - start_ts:.2f}s.")
 
         if not len(all_entities_data) and not len(all_relationships_data):
-            logging.warning(
-                "Didn't extract any entities and relationships, maybe your LLM is not working"
-            )
+            logging.warning("Didn't extract any entities and relationships, maybe your LLM is not working")
 
         if not len(all_entities_data):
             logging.warning("Didn't extract any entities")
@@ -167,15 +163,11 @@ class Extractor:
         if not entities:
             return
         entity_type = sorted(
-            Counter(
-                [dp["entity_type"] for dp in entities]
-            ).items(),
+            Counter([dp["entity_type"] for dp in entities]).items(),
             key=lambda x: x[1],
             reverse=True,
         )[0][0]
-        description = GRAPH_FIELD_SEP.join(
-            sorted(set([dp["description"] for dp in entities]))
-        )
+        description = GRAPH_FIELD_SEP.join(sorted(set([dp["description"] for dp in entities])))
         already_source_ids = flat_uniq_list(entities, "source_id")
         description = await self._handle_entity_relation_summary(entity_name, description)
         node_data = dict(
@@ -186,13 +178,7 @@ class Extractor:
         node_data["entity_name"] = entity_name
         all_relationships_data.append(node_data)
 
-    async def _merge_edges(
-            self,
-            src_id: str,
-            tgt_id: str,
-            edges_data: list[dict],
-            all_relationships_data=None
-    ):
+    async def _merge_edges(self, src_id: str, tgt_id: str, edges_data: list[dict], all_relationships_data=None):
         if not edges_data:
             return
         weight = sum([edge["weight"] for edge in edges_data])
@@ -200,14 +186,7 @@ class Extractor:
         description = await self._handle_entity_relation_summary(f"{src_id} -> {tgt_id}", description)
         keywords = flat_uniq_list(edges_data, "keywords")
         source_id = flat_uniq_list(edges_data, "source_id")
-        edge_data = dict(
-            src_id=src_id,
-            tgt_id=tgt_id,
-            description=description,
-            keywords=keywords,
-            weight=weight,
-            source_id=source_id
-        )
+        edge_data = dict(src_id=src_id, tgt_id=tgt_id, description=description, keywords=keywords, weight=weight, source_id=source_id)
         all_relationships_data.append(edge_data)
 
     async def _merge_graph_nodes(self, graph: nx.Graph, nodes: list[str], change: GraphChange):
@@ -243,14 +222,10 @@ class Extractor:
         node0_attrs["description"] = await self._handle_entity_relation_summary(nodes[0], node0_attrs["description"])
         graph.nodes[nodes[0]].update(node0_attrs)
 
-    async def _handle_entity_relation_summary(
-            self,
-            entity_or_relation_name: str,
-            description: str
-    ) -> str:
+    async def _handle_entity_relation_summary(self, entity_or_relation_name: str, description: str) -> str:
         summary_max_tokens = 512
         use_description = truncate(description, summary_max_tokens)
-        description_list=use_description.split(GRAPH_FIELD_SEP),
+        description_list = (use_description.split(GRAPH_FIELD_SEP),)
         if len(description_list) <= 12:
             return use_description
         prompt_template = SUMMARIZE_DESCRIPTIONS_PROMPT